+++ Common/IDigitCalculator.cs
@@ -4,7 +4,7 @@
     {
         T CalculateDigit(T value, T carry);
         T CalculateCarry(T value);
-        T Cast(uint value);
+        T Cast(ulong value);
         bool IsDigitGreaterThanZero(T value);
         T CalculateSum(T a, T b, T c);
         T CalculateSum(T a, T b);
